[R1378] merging DocumentVocabulary with Vocabulary
