+++ src/main/java/cz/cvut/kbss/termit/model/Vocabulary.java
@@ -20,6 +20,7 @@ import cz.cvut.kbss.jsonld.annotation.JsonLdAttributeOrder;
 import cz.cvut.kbss.termit.exception.TermItException;
 import cz.cvut.kbss.termit.model.changetracking.Audited;
 
+import cz.cvut.kbss.termit.model.resource.Document;
 import javax.validation.constraints.NotBlank;
 import java.io.Serializable;
 import java.lang.reflect.Field;
@@ -41,6 +42,9 @@ public class Vocabulary extends Asset<String> implements Serializable {
 
     @OWLAnnotationProperty(iri = DC.Terms.DESCRIPTION)
     private String description;
+
+    @OWLObjectProperty(iri = cz.cvut.kbss.termit.util.Vocabulary.s_p_popisuje_dokument, cascade = CascadeType.PERSIST, fetch = FetchType.EAGER)
+    private Document document;
 
     @ParticipationConstraints(nonEmpty = true)
     @OWLObjectProperty(iri = cz.cvut.kbss.termit.util.Vocabulary.s_p_ma_glosar, cascade = CascadeType.PERSIST,
@@ -74,6 +78,14 @@ public class Vocabulary extends Asset<String> implements Serializable {
 
     public void setDescription(String description) {
         this.description = description;
+    }
+
+    public Document getDocument() {
+        return document;
+    }
+
+    public void setDocument(Document document) {
+        this.document = document;
     }
 
     public Glossary getGlossary() {
