+++ src/main/java/cz/cvut/kbss/termit/persistence/DescriptorFactory.java
@@ -19,7 +19,6 @@ import cz.cvut.kbss.jopa.model.descriptors.Descriptor;
 import cz.cvut.kbss.jopa.model.descriptors.EntityDescriptor;
 import cz.cvut.kbss.jopa.model.descriptors.FieldDescriptor;
 import cz.cvut.kbss.jopa.model.metamodel.FieldSpecification;
-import cz.cvut.kbss.termit.model.DocumentVocabulary;
 import cz.cvut.kbss.termit.model.Glossary;
 import cz.cvut.kbss.termit.model.Term;
 import cz.cvut.kbss.termit.model.Vocabulary;
@@ -86,14 +85,14 @@ public class DescriptorFactory {
         Objects.requireNonNull(vocabularyUri);
         final EntityDescriptor descriptor = assetDescriptor(vocabularyUri);
         descriptor.addAttributeDescriptor(fieldSpec(Vocabulary.class, "glossary"), glossaryDescriptor(vocabularyUri));
-        descriptor.addAttributeDescriptor(fieldSpec(DocumentVocabulary.class, "document"),
+        descriptor.addAttributeDescriptor(fieldSpec(Vocabulary.class, "document"),
                 documentDescriptor(vocabularyUri));
         return descriptor;
     }
 
     /**
      * Creates a JOPA descriptor for a {@link Document} related to the specified vocabulary (presumably a {@link
-     * DocumentVocabulary}).
+     * Vocabulary}).
      * <p>
      * This means that the context of the Document (and all its relevant attributes) is given by the specified
      * vocabulary's IRI.
@@ -110,7 +109,7 @@ public class DescriptorFactory {
 
     /**
      * Creates a JOPA descriptor for a {@link Document} related to a vocabulary with the specified identifier
-     * (presumably of a {@link DocumentVocabulary}).
+     * (presumably of a {@link Vocabulary}).
      * <p>
      * This means that the context of the Document (and all its relevant attributes) is given by the specified IRI.
      * <p>
