+++ src/test/java/cz/cvut/kbss/termit/persistence/dao/VocabularyDaoTest.java
@@ -118,7 +118,7 @@ class VocabularyDaoTest extends BaseDaoTestRunner {
 
     @Test
     void updateWorksCorrectlyInContextsForDocumentVocabulary() {
-        final DocumentVocabulary vocabulary = new DocumentVocabulary();
+        final Vocabulary vocabulary = new Vocabulary();
         vocabulary.setUri(Generator.generateUri());
         vocabulary.setLabel("test-vocabulary");
         vocabulary.setGlossary(new Glossary());
